Split utils to align with Unreal; Removed unused imports
