+++ doc/developing/api_references/ftrack_connect_pipeline_maya.utils.rst
@@ -5,3 +5,13 @@ ftrack\_connect\_pipeline\_maya.utils package
    :members:
    :undoc-members:
    :show-inheritance:
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 1
+
+   ftrack_connect_pipeline_maya.plugin.bootstrap
+   ftrack_connect_pipeline_maya.plugin.file
+   ftrack_connect_pipeline_maya.plugin.node
