+++ resource/plugins/python/publisher/exporters/maya_native_publisher_exporter.py
@@ -6,7 +6,7 @@ import os
 
 import maya.cmds as cmds
 
-from ftrack_connect_pipeline_maya.utils import custom_commands as maya_utils
+from ftrack_connect_pipeline_maya import utils as maya_utils
 from ftrack_connect_pipeline_maya import plugin
 import ftrack_api
 
@@ -63,7 +63,7 @@ class MayaNativePublisherExporterPlugin(plugin.MayaPublisherExporterPlugin):
                 # Scene is not saved, save it first. Should have been taken
                 # care of by scene collector.
                 self.logger.warning('Maya not saved, saving locally..')
-                save_path, message = maya_utils.save(
+                save_path, message = maya_utils.save_file(
                     context_data['context_id'], self.session
                 )
                 if not message is None:
