+++ source/ftrack_connect_pipeline_maya/client/asset_manager.py
@@ -9,7 +9,6 @@ import ftrack_connect_pipeline_qt.constants as qt_constants
 import ftrack_connect_pipeline_maya.constants as maya_constants
 
 from maya.app.general.mayaMixin import MayaQWidgetDockableMixin
-from ftrack_connect_pipeline_maya.utils.custom_commands import get_main_window
 
 
 class MayaQtAssetManagerClientWidget(QtAssetManagerClientWidget):
