+++ source/ftrack_connect_pipeline_maya/client/log_viewer.py
@@ -5,8 +5,6 @@ from Qt import QtWidgets, QtCore
 
 from ftrack_connect_pipeline_qt.client import log_viewer
 
-from ftrack_connect_pipeline_maya.utils.custom_commands import get_main_window
-
 
 class MayaQtLogViewerClientWidget(log_viewer.QtLogViewerClientWidget):
     '''Maya log viewer dialog'''
