+++ source/ftrack_connect_pipeline_maya/client/open.py
@@ -4,17 +4,16 @@
 from Qt import QtWidgets, QtCore
 
 from ftrack_connect_pipeline_qt.client import open
-import ftrack_connect_pipeline.constants as constants
-import ftrack_connect_pipeline_maya.constants as maya_constants
-from ftrack_connect_pipeline_maya.utils.custom_commands import get_main_window
+import ftrack_connect_pipeline.constants as core_constants
 from ftrack_connect_pipeline_qt import constants as qt_constants
+import ftrack_connect_pipeline_maya.constants as maya_constants
 
 
 class MayaQtOpenerClientWidget(open.QtOpenerClientWidget):
     '''Open dialog and client'''
 
     ui_types = [
-        constants.UI_TYPE,
+        core_constants.UI_TYPE,
         qt_constants.UI_TYPE,
         maya_constants.UI_TYPE,
     ]
