+++ source/ftrack_connect_pipeline_maya/client/save.py
@@ -1,7 +1,7 @@
 # :coding: utf-8
 # :copyright: Copyright (c) 2014-2022 ftrack
 from ftrack_connect_pipeline_qt.client.save import QtSaveClientWidget
-from ftrack_connect_pipeline_maya.utils import custom_commands as maya_utils
+from ftrack_connect_pipeline_maya import utils as maya_utils
 
 
 class MayaQtSaveClientWidget(QtSaveClientWidget):
