+++ source/ftrack_connect_pipeline_maya/plugin/open/importer.py
@@ -10,9 +10,8 @@ from ftrack_connect_pipeline_maya.plugin import (
     MayaBasePluginWidget,
 )
 
-from ftrack_connect_pipeline_maya.utils import custom_commands as maya_utils
+from ftrack_connect_pipeline_maya import utils as maya_utils
 from ftrack_connect_pipeline_maya.constants.asset import modes as load_const
-from ftrack_connect_pipeline_maya.constants import asset as asset_const
 
 
 class MayaOpenerImporterPlugin(plugin.OpenerImporterPlugin, MayaBasePlugin):
