+++ source/ftrack_connect_pipeline_maya/plugin/publish/finalizer.py
@@ -9,7 +9,7 @@ from ftrack_connect_pipeline_maya.plugin import (
 )
 
 import maya.cmds as cmds
-from ftrack_connect_pipeline_maya.utils import custom_commands as maya_utils
+from ftrack_connect_pipeline_maya import utils as maya_utils
 from ftrack_connect_pipeline_maya.constants import asset as asset_const
